Create README.md
This is a Nokia example application that acts as a client for the [Nokia Beta Labs](http://betalabs.nokia.com/). Developed with [Qt Quick Components](http://doc.qt.nokia.com/qt-components-symbian-1.1/index.html), this example demonstrates [[How to use the In-App Advertising API]] and how to elegantly integrate ads to your application. 

[[Image(wiki:WikiStart:blc2-latest-page.png, 200px)]]
[[Image(wiki:WikiStart:blc2-details-page.png, 200px)]]
[[Image(wiki:WikiStart:blc2-search-page-with-ad.png, 200px)]]

### Compatibility
* Symbian (Qt 4.7.4, Qt Quick Components 1.1 and Qt Mobility 1.2.1)

Tested on:
* Nokia 701
* Nokia C7-00
* Nokia E7-00
* Nokia N8-00
* Nokia X7-00

Developed with Qt SDK 1.2.